+++ packages/ranger-vite-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ranger-theme/vite-config
 
+## 0.2.0
+
+### Minor Changes
+
+- add https option web dev server
+
 ## 0.1.0
 
 ### Minor Changes